Move InsertResponse
Move class definition to its own file, no longer nested under Table.
