+++ lib/gcloud/bigquery/table.rb
@@ -17,6 +17,7 @@ require "gcloud/bigquery/view"
 require "gcloud/bigquery/data"
 require "gcloud/bigquery/table/list"
 require "gcloud/bigquery/errors"
+require "gcloud/bigquery/insert_response"
 require "gcloud/upload"
 
 module Gcloud
@@ -820,68 +821,6 @@ module Gcloud
 
       def data_complete?
         !@gapi["creationTime"].nil?
-      end
-
-      ##
-      # InsertResponse
-      class InsertResponse
-        def initialize rows, gapi #:nodoc:
-          @rows = rows
-          @gapi = gapi
-        end
-
-        def success?
-          error_count.zero?
-        end
-
-        def insert_count
-          @insert_count ||= @rows.count - error_count
-        end
-
-        def error_count
-          @error_count ||= Array(@gapi["insertErrors"]).count
-        end
-
-        def insert_errors
-          @insert_errors ||= begin
-            Array(@gapi["insertErrors"]).map do |ie|
-              row = @rows[ie["index"]]
-              errors = ie["errors"]
-              InsertError.new row, errors
-            end
-          end
-        end
-
-        def error_rows
-          @error_rows ||= begin
-            Array(@gapi["insertErrors"]).map do |ie|
-              @rows[ie["index"]]
-            end
-          end
-        end
-
-        def errors_for row
-          ie = insert_errors.detect { |e| e.row == row }
-          return ie.errors if ie
-          []
-        end
-
-        def self.from_gapi rows, gapi #:nodoc:
-          gapi = gapi.to_hash if gapi.respond_to? :to_hash
-          new rows, gapi
-        end
-
-        ##
-        # InsertError
-        class InsertError
-          attr_reader :row
-          attr_reader :errors
-
-          def initialize row, errors #:nodoc:
-            @row = row
-            @errors = errors
-          end
-        end
       end
     end
   end
